% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vélo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution hebdomadaire des trajets}
\usage{
calcul_distribution_semaine(trajet, filtre = TRUE)
}
\arguments{
\item{filtre}{Booléen. Si TRUE, les anomalies sont filtrées avant calcul.
Si FALSE, le jeu de données est utilisé tel quel.}
}
\value{
Un data frame avec le jour de la semaine et le nombre
total de trajets.
}
\description{
Cette fonction calcule le nombre total de trajets
par jour de la semaine.
}
