% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vélo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles distinctes}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{Un data frame contenant les données de trajets.}
}
\value{
Un entier correspondant au nombre de boucles distinctes.
}
\description{
Cette fonction compte le nombre de valeurs distinctes
dans la colonne \verb{Numéro de boucle}.
}
