% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vélo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre total de trajets}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{Un data frame contenant les données de trajets.}
}
\value{
Un nombre correspondant au total des trajets.
}
\description{
Cette fonction calcule la somme de la variable \code{Total}.
}
