% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnee.R
\docType{data}
\name{donnee}
\alias{donnee}
\title{Jeu de données df_velo}
\format{
A data frame with XX rows and XX variables:
\describe{
\item{var1}{Description de la variable 1}
\item{var2}{Description de la variable 2}
\item{var3}{Description de la variable 3}
}
}
\source{
Fichier de données importé dans le package.
}
\usage{
donnee
}
\description{
Données utilisées pour l'entraînement et l'apprentissage de la création
d'un package R.
}
\keyword{datasets}
