% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vélo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les trajets sans anomalie détectée}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data frame contenant les données de trajets.}
}
\value{
Un data frame filtré sans les trajets signalés comme anormaux.
}
\description{
Cette fonction conserve uniquement les lignes pour lesquelles
la probabilité de présence d'anomalies est manquante.
}
