% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vélo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets selon un ou plusieurs numéros del boucleessss}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{Un data frame contenant les données de trajets.}

\item{boucle}{Un vecteur de caractères contenant les numéros de boucle à conserver.}
}
\value{
Un data frame filtré selon les boucles sélectionnées.
}
\description{
Cette fonction filtre un jeu de données de trajets pour ne conserver
que les lignes correspondant aux numéros de boucle sélectionnés.
}
