% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vélo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Représenter la distribution hebdomadaire des trajets}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data frame contenant les données de trajets.}
}
\value{
Un graphique ggplot2.
}
\description{
Cette fonction produit un diagramme en barres du nombre total
de trajets par jour de la semaine, après filtrage des anomalies.
}
