% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vélo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet ayant le plus grand total}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data frame contenant les données de trajets.}
}
\value{
Un data frame contenant la boucle de comptage, le jour
et le total du trajet maximal.
}
\description{
Cette fonction retourne la ligne correspondant au trajet
ayant la valeur maximale de \code{Total}.
}
